network: Introduce IO context struct
Add a common IO context structure that is shared between the
iio_context_pdata struct and iio_device_pdata struct. This will allow to
pass additional state besides the socket to IO operations.

For now the IO context only the socket file descriptor, but additional
state will be added in the following commits.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>